assistant-server/requirements.txt
numpy=-1.23.5
pysampi==0.39.1
PyAudio==0.2.13
aiohttp==3.8.4
openai==0.27.4
openai_whisper==20230314
python-dotenv==1.0.0
aiohttp==3.8.4
gunicorn==21.2.0
Flask==2.2.3
requests==2.28.2
python-dotenv==1.0.0